+++ simplyblock_web/api/v2/cluster/backup.py
@@ -1,19 +1,19 @@
-from typing import Annotated, List, Optional
+from typing import List, Optional
 from uuid import UUID
 
-from fastapi import APIRouter, Depends, HTTPException, Query, Response
+from fastapi import APIRouter, HTTPException, Query, Response
 from pydantic import BaseModel
 
 from simplyblock_core.db_controller import DBController
 from simplyblock_core.controllers import backup_controller
-from simplyblock_core.models.backup import BackupPolicy
 from simplyblock_core.models.cluster import Cluster as ClusterModel
 from simplyblock_core.models.lvol_model import LVol
 
-from .cluster import Cluster
-from .dtos import BackupDTO, BackupPolicyDTO
+from .._dependencies import Cluster, Policy
+from .._dtos import BackupDTO, BackupPolicyDTO
 
-api = APIRouter(prefix='/backups')
+
+api = APIRouter()
 db = DBController()
 
 
@@ -135,7 +135,7 @@ def delete_backups(cluster: Cluster, volume_id: UUID) -> Response:
 
 # Backup policies
 
-policy_api = APIRouter(prefix='/backup-policies')
+policy_api = APIRouter()
 
 
 @policy_api.get('/', name='clusters:backup-policies:list')
@@ -161,19 +161,6 @@ def create_policy(cluster: Cluster, parameters: _PolicyCreateParams) -> Response
     if error:
         raise HTTPException(400, error)
     return Response(status_code=201, headers={'X-Policy-Id': policy_id})
-
-
-def _lookup_backup_policy(policy_id: UUID, cluster: Cluster) -> BackupPolicy:
-    try:
-        policy = db.get_backup_policy_by_id(str(policy_id))
-    except KeyError as e:
-        raise HTTPException(404, str(e))
-    if policy.cluster_id != cluster.get_id():
-        raise HTTPException(404, f'BackupPolicy {policy_id} not found')
-    return policy
-
-
-Policy = Annotated[BackupPolicy, Depends(_lookup_backup_policy)]
 
 
 def _validate_attachment_target(target_type: str, target_id: str, cluster: ClusterModel) -> None:
@@ -219,3 +206,6 @@ def detach_policy(cluster: Cluster, policy: Policy, parameters: _AttachParams) -
     if error:
         raise HTTPException(400, error)
     return Response(status_code=204)
+
+
+api.include_router(policy_api, prefix='/backup-policies')
