+++ simplyblock_web/api/v2/cluster/task.py
@@ -1,14 +1,15 @@
-from typing import Annotated, List
-from uuid import UUID
+from typing import List
 
-from fastapi import APIRouter, Depends, HTTPException
+from fastapi import APIRouter
 
 from simplyblock_core.db_controller import DBController
 from simplyblock_core.models.job_schedule import JobSchedule
-from .cluster import Cluster
-from .dtos import TaskDTO
 
-api = APIRouter(prefix='/tasks')
+from .._dependencies import Cluster, Task
+from .._dtos import TaskDTO
+
+
+api = APIRouter()
 db = DBController()
 
 
@@ -26,18 +27,9 @@ def list(cluster: Cluster) -> List[TaskDTO]:
 instance_api = APIRouter(prefix='/{task_id}')
 
 
-def _lookup_task(task_id: UUID, cluster: Cluster) -> JobSchedule:
-    task = db.get_task_by_id(str(task_id))
-    if task is None:
-        raise HTTPException(404, 'Task does not exist')
-    if task.cluster_id != cluster.get_id():
-        raise HTTPException(404, 'Task does not exist')
-    return task
-
-
-Task = Annotated[JobSchedule, Depends(_lookup_task)]
-
-
 @instance_api.get('/', name='clusters:tasks:detail')
 def get(cluster: Cluster, task: Task) -> TaskDTO:
     return TaskDTO.from_model(task)
+
+
+api.include_router(instance_api)
